Crate dotfiles_core

The core of Dotfiles-rs is basically a set of directives that can build executable actions from configuration sources (Yaml is the only supported configuration source so far).

Directives are responsible for parsing configuration from the configuration source, applying default settings if necessary, and using these to build actions that can be executed.

The base traits and classes for these building blocks can be found in the action and directive modules. The yaml_util module is a set of helper functions for Directives to parse configuration.

Finally all concrete actions and directives can be found in the following modules:

create: Create a new directory

Re-exports

pub use action::Action;

pub use directive::Directive;

pub use directive::Setting;

pub use directive::Settings;

Modules

action

This module contains the base trait for all Actions.

brew

This module contains the BrewAction and BrewDirective

create

This module contains the CreateAction and CreateDirective classes that are used by the runtime to create a new directory.

directive

This module contains the base trait for all Directive and all necessary conveniences to allow for user-configuration of directive defaults.

homebrew_install

This module contains the HomebrewInstallAction

link

This module contains the LinkAction and LinkDirective classes that are used by the runtime to create a symlink.

ohmyzsh_install

This module contains the OhMyZshInstallAction

yaml_util

Module that defines helper functions to process YAML configuration sources.