Module dotfiles_core::action

This module contains the base trait for all Actions.

Constants

SKIP_IN_CI_SETTING

Skip this whole action in CI environments.

Traits

Action

An action to be run by a the dotfiles runtime.

ActionParser

Trait to parse a specific action type from StrictYaml.

ConditionalAction

Trait for actions to be skippable under certain conditions.

Functions

is_running_in_ci

Whether the execution environment is presumed to be CI