dotfiles_core::action
Function is_running_in_ci

pub fn is_running_in_ci() -> bool

Whether the execution environment is presumed to be CI

The presence of any of the following Environment Variables is assumed to mean that this action is running in a CI Environment:

TF_BUILD

BUILDKITE

BUILD_ID

CI

CIRCLECI

CIRRUS_CI

CODEBUILD_BUILD_ID

GITLAB_CI

GITHUB_ACTIONS

HEROKU_TEST_RUN_ID

TEAMCITY_VERSION

TRAVIS