Module dotfiles_core::create::directive

This module defines CreateDirective.

Structs

CreateDirective

A directive that can build CreateActions to create directories in the filesystem.

Constants

DIRECTIVE_NAME

Constant for the name of the create directive.

DIR_SETTING

Constant for the name of the directory Setting which forces to create all parent directories if necessary.

FORCE_SETTING

Constant for the name of the force Setting which forces to create all parent directories if necessary.

Functions

init_directive_data

Initializes the default configuration for the CreateDirective

new_native_create_directive

Constructs a new CreateDirective using the real filesystem