dotfiles_core::error

Enum ErrorType

pub enum ErrorType {
    ExecutionError {
        io_error: Option<Error>,
        exit_status: Option<ExitStatus>,
    },
    FileSystemError {
        fs_error: Error,
    },
    InconsistentConfigurationError,
    IncompleteConfigurationError {
        missing_field: String,
    },
    YamlParseError {
        scan_error: ScanError,
    },
    UnexpectedYamlTypeError {
        encountered: StrictYaml,
        expected: StrictYaml,
    },
    CoreError,
    TestingErrorActionSucceedsWhenItShouldFail,
}

A collection of types of errors that may occur while parsing or executing actions

Variants

ExecutionError

An error occurred while running a command necessary for executing an action

Fields

io_error: Option<Error>

If the command could not execute for some reason the underlying io Error will be saved here

exit_status: Option<ExitStatus>

If the command attempted to execute but failed for some reason, the underlying ExitStatus will be saved here.

FileSystemError

A filesystem error that was encountered while either reading configuration or executing a filesystem related action

Fields

fs_error: Error

The underlying filesystem error.

InconsistentConfigurationError

The configuration file is inconsistent with itself or with that dotfiles supports.

IncompleteConfigurationError

The configuration is missing a required field

Fields

missing_field: String

Name of the field missing in the configuration

YamlParseError

An error that occurred while parsing the StrictYaml file

Fields

scan_error: ScanError

The underlying scan error

UnexpectedYamlTypeError

Received an StrictYaml object of an unexpected type

Fields

encountered: StrictYaml

What we got instead of the expected type.

expected: StrictYaml

An example of what we expected.

CoreError

A core logic error for Dotfiles-rs

TestingErrorActionSucceedsWhenItShouldFail

An error only for testing, the action that should fail actually succeeds!