dotfiles_core::exec_wrapper

Function execute_commands

pub fn execute_commands(
    cmds: Vec<Exec>,
    popen_error_message: &str,
    error_while_running_message: &str,
) -> Result<(), DotfilesError>

Executes the cmd and waits for it to finish.

If the execution returns a [subprocess::PopenError] then it uses the popen_error_message for the message in a DotfilesError.

If the execution finishes but in an error state, then it uses the error_while_running_message instead.