Struct dotfiles_core::link::action::LinkAction

pub struct LinkAction<'a, F: FileSystem + UnixFileSystem> { /* fields omitted */ }

LinkAction creates a new symlink path that points to target.

It is equivalent to running ln -s <target> <path>

Implementations

impl<'a, F: FileSystem + UnixFileSystem> LinkAction<'a, F>

pub fn new(
fs: &'a F,
path: String,
target: String,
context_settings: &Settings,
defaults: &Settings
) -> LinkAction<'a, F>

Constructs a new LinkAction

pub fn path(&self) -> &str

Path of the new symlink

pub fn target(&self) -> &str

Path that the symlink points to.

pub fn relink(&self) -> bool

Force to re-create the symlink if it exists already

pub fn force(&self) -> bool

Force to replace an existing file or directory when executed.

pub fn create_parent_dirs(&self) -> bool

Create all parent directories if they do not exist already

pub fn ignore_missing_target(&self) -> bool

Succeed even if target doesn’t point to an existing file or dir.

pub fn relative(&self) -> bool

Allow relative linking. TODO: actually implement relative linking.

pub fn resolve_symlink_target(&self) -> bool

If the target is another symlink, resolve the ultimate concrete file or directory that it points to and make it the target