Module dotfiles_core::link::directive

This module defines LinkDirective.

Structs

LinkDirective

A directive that can build LinkActions to create directories in the filesystem.

Constants

CREATE_PARENT_DIRS_SETTING

Create parent dirs if they don’t exist

DIRECTIVE_NAME

Name of the link directive

FORCE_SETTING

Force setting, replaces any other file or directory

IGNORE_MISSING_TARGET_SETTING

Create the symlink even if the target file does not exist

PATH_SETTING

Path setting (path of the symlink)

RELATIVE_SETTING

TODO: Allow relative symlinks, if false any relative symlinks cause a failure.

RELINK_SETTING

Relink setting, if true the action relinks an existing symlink (applies if force is false)

RESOLVE_SYMLINK_TARGET_SETTING

Resolves the target if it is a symlink and uses the final target file as the target.

TARGET_SETTING

Target setting (path to the file the symlink points to)

Functions

init_directive_data

Initialize the defaults for the LinkDirective.

new_native_link_directive

Create a new link directive using the native filesystem