Module dotfiles_core::path

Contains helpful functions to deal with paths in the context of parsing them for dotfiles configs.

Functions

convert_path_to_absolute

Converts a file path to absolute if it is relative. If current_dir is provided it uses it for the base dir, otherwise it relies on std::env::current_dir()

process_home_dir_in_path

Checks for ~ and replaces it with a home directory if necessary.