dotfiles_core

Module settings

This module contains the definition of a setting and code to parse them.

Enums

Setting

Represents a value for a setting

Functions

initialize_settings_object

Returns a Settings object from an array as a bit of syntactic sugar

parse_setting

Parse a setting from StrictYaml given a particular setting type.

Type Aliases

Settings

The Settings object is a hashmap of option names to a default value