Module dotfiles_core::yaml_util

Module that defines helper functions to process YAML configuration sources.

Functions

get_boolean_setting

Gets a boolean value for the setting named name.

get_boolean_setting_from_yaml_or_defaults

Gets a Boolean value from YAML or defaults.

get_int_setting

Gets a Int value for the setting named name.

get_integer_setting_from_yaml_or_defaults

Gets a Integer value from YAML or defaults.

get_setting

Gets a String value for the setting named name.

get_string_array

Gets a native Vec<String> from a Yaml::Array. It errors out if the passed yaml is not an array or if not all the items in the array are plain Yaml Strings

get_string_content_or_keyed_value

Gets the content of this YAML node or the value for a specific key in it.

get_string_setting

Gets a String value for the setting named name.

get_string_setting_from_yaml_or_defaults

Gets a String value from YAML or defaults.