Function dotfiles_core::yaml_util::get_string_setting_from_yaml_or_context

pub fn get_string_setting_from_yaml_or_context(
    name: &str,
    yaml: &StrictYaml,
    context_settings: &Settings,
    directive_defaults: &Settings
) -> Result<String, DotfilesError>

Gets a String value from YAML or context.

First itb tries to find a value in yaml, if it can’t find one then it falls back to context_settings or finally default_settings.

Errors

Found a setting in YAML but that couldn’t be parsed.

Didn’t find a setting matching this name anywhere