Function dotfiles_core::yaml_util::map_yaml_array

pub fn map_yaml_array<T, F>(
    yaml_array: &StrictYaml,
    process: F
) -> Result<Vec<T>, DotfilesError>where
    F: FnMut(&StrictYaml) -> Result<T, DotfilesError>,

Calls a processing function on all elements of an array, will fail if any of the elements fail to process.

Errors

Any error that happens in the processing function.

ErrorType::UnexpectedYamlTypeError if the yaml passed is not an array