dotfiles_core::yaml_util
Function process_yaml_hash_until_first_err

pub fn process_yaml_hash_until_first_err<F>(
    yaml_hash: &StrictYaml,
    process_function: F,
) -> Result<(), DotfilesError>
where
    F: FnMut(String, &StrictYaml) -> Result<(), DotfilesError>,

Executes the process_function on each of the items in the yaml_hash. The yaml hash is assumed to be string keyed. It stops execution if any of the process functions returns an Error, and returns said error.