dotfiles_core

Module directive

This module contains the base trait for all Directive and all necessary conveniences to allow for user-configuration of directive defaults.

Structs

DirectiveData

A struct that contains the default settings for a Directive and the name it takes in configuration sources. The name must be unique.

Traits

Directive

A parser for action steps, each directive represents a type of Action.

HasDirectiveData

A trait for all directives, it is shared between crate::action::ActionParser and Directive