Trait dotfiles_core::directive::Directive

pub trait Directive<'a>: HasDirectiveData<'a> {
    fn build_action_list(
        &'a self,
        settings: &Settings,
        yaml: &Yaml
    ) -> Result<Vec<Box<dyn Action<'a> + 'a>>, DotfilesError>;

    fn name(&'a self) -> &'a str { ... }
    fn defaults(&'a self) -> &'a Settings { ... }
    fn get_setting_from_yaml_hash_or_from_context(
        &'a self,
        name: &str,
        yaml: &Yaml,
        context_settings: &Settings
    ) -> Result<Setting, DotfilesError> { ... }
    fn get_setting_from_yaml_hash(
        &'a self,
        name: &str,
        yaml: &Yaml
    ) -> Result<Setting, DotfilesError> { ... }
    fn parse_setting_value(
        &'a self,
        name: &str,
        yaml: &Yaml
    ) -> Result<Setting, DotfilesError> { ... }
    fn parse_context_defaults(
        &'a self,
        yaml_settings: &Yaml
    ) -> Result<Settings, DotfilesError> { ... }
}

A parser for action steps, each directive represents a type of Action.

Required Methods

fn build_action_list(
&'a self,
settings: &Settings,
yaml: &Yaml
) -> Result<Vec<Box<dyn Action<'a> + 'a>>, DotfilesError>

Builds a list of actions for this directive from a Yaml configuration object and a set of default settings.

Returns an Error containing a human readable string in case there was an issue building the actions.

Provided Methods

fn name(&'a self) -> &'a str

Returns the name of the directive.

fn defaults(&'a self) -> &'a Settings

Returns the default settings as configured.

fn get_setting_from_yaml_hash_or_from_context(
&'a self,
name: &str,
yaml: &Yaml,
context_settings: &Settings
) -> Result<Setting, DotfilesError>

Parse a particular setting with its correct type from yaml, fall back to context settings or directive defaults if not found in yaml. Returns error if there is any kind of parsing or typing error

fn get_setting_from_yaml_hash(
&'a self,
name: &str,
yaml: &Yaml
) -> Result<Setting, DotfilesError>

Parses an individual setting named name from a yaml hash using the type stored in DirectiveData.setting_types.

fn parse_setting_value(
&'a self,
name: &str,
yaml: &Yaml
) -> Result<Setting, DotfilesError>

DirectiveData.setting_types.

fn parse_context_defaults(
&'a self,
yaml_settings: &Yaml
) -> Result<Settings, DotfilesError>

Parses all settings for this directive from Yaml, checking the types correspond to what’s stored in DirectiveData.setting_types