Trait dotfiles_core::directive::Directive

pub trait Directive<'a> {
    fn name(&self) -> &str;
    fn defaults(&self) -> &Settings;
    fn build_action(
        &'a self, 
        settings: &Settings, 
        yaml: &Yaml
    ) -> Result<Vec<Box<dyn Action<'a> + 'a>>, String>;
}

A parser for action steps, each directive represents a type of Action.

Required Methods

fn name(&self) -> &str

Returns the name of the directive.

fn defaults(&self) -> &Settings

Returns the defaults settings as configured.

fn build_action(
&'a self,
settings: &Settings,
yaml: &Yaml
) -> Result<Vec<Box<dyn Action<'a> + 'a>>, String>

Builds an action from a Yaml configuration source and a set of default settings.

Returns an Error containing a human readable string in case there was an issue building the action.