Enum dotfiles_core::error::ErrorType

pub enum ErrorType {
    ExecutionError {
        popen_error: Option<PopenError>,
        exit_status: Option<ExitStatus>,
    },
    FileSystemError {
        fs_error: Error,
    },
    InconsistentConfigurationError,
    IncompleteConfigurationError {
        missing_field: String,
    },
    YamlParseError {
        scan_error: ScanError,
    },
    UnexpectedYamlTypeError {
        encountered: StrictYaml,
        expected: StrictYaml,
    },
    CoreError,
    TestingErrorActionSucceedsWhenItShouldFail,
}

A collection of types of errors that may occur while parsing or executing actions

Variants

ExecutionError

Fields

popen_error: Option<PopenError>

If the command could not execute for some reason the underlying Popen Error will be saved here

exit_status: Option<ExitStatus>

If the command attempted to execute but failed for some reason, the underlying ExitStatus will be saved here.

An error occurred while running a command necessary for executing an action

FileSystemError

Fields

fs_error: Error

The underlying filesystem error.

A filesystem error that was encountered while either reading configuration or executing a filesystem related action

InconsistentConfigurationError

The configuration file is inconsistent with itself or with that dotfiles supports.

IncompleteConfigurationError

Fields

missing_field: String

Name of the field missing in the configuration

The configuration is missing a required field

YamlParseError

Fields

scan_error: ScanError

The underlying scan error

An error that occurred while parsing the StrictYaml file

UnexpectedYamlTypeError

Fields

encountered: StrictYaml

What we got instead of the expected type.

expected: StrictYaml

An example of what we expected.

Received an StrictYaml object of an unexpected type

CoreError

A core logic error for Dotfiles-rs

TestingErrorActionSucceedsWhenItShouldFail

An error only for testing, the action that should fail actually succeeds!