Function dotfiles_core::path::convert_path_to_absolute

pub fn convert_path_to_absolute(
    file_name: &Path,
    current_dir: Option<&Path>
) -> Result<PathBuf, DotfilesError>

Converts a file path to absolute if it is relative. If current_dir is provided it uses it for the base dir, otherwise it relies on std::env::current_dir()